bug: Use `from cadquery.vis import show`
And also remote `try except` as it was always throwing.

The reason is the the `from cadquery import show` isn't working. I
asked the bot:
  https://chatgpt.com/share/682a368c-960c-800c-af7b-ee5356579829

And it suggested a difference in configuration and likely
in what's in .direnv as it contians a copy of the submodule and I'm
editing the show command.

Finally, I told the bot that dev clone, cq-sweep-1, worked and the
test clone, cq-sweep, wasn't. I lied as both are now not working, the
reason is I deleted .direnv on the cq-sweep-1 and did another
`direnv allow`, thus leading to my guess it's was something in .direnv.

Anyway removing the exception handling and using importing show
from cadquery.vis :)